Initial Elmo Setup
This initial iteration of the ELMO setup script allows the user to
specify the generator to create flows on, bitrate, number of subports,
and the number of flows per subport.  Each flow will automatically be
generated with a unique double-vlan tag.  The limiting factor of flow
generation is the number of flows ELMO2 can sustain per generator.  This
script is reliant on the ELMO test automation package to run.

The flows generated by this script are compliant with the packet
formatting specifications of DPDK v2.0's qos_sched application.